Handle variable attributes:
Discovered by Maciej Sokolowski.

This is a tricky one because PPI doesn't parse the document tree
correctly. It thinks the variable attribute is a new statement and
not part of the existing statement.

The solution, however, is simple. We remove spaces and then just
access the next token. If it's in the same statement, it will move
to the next sibling. If it's in a new statement (as PPI understands
this), it will move to the new statement.

Then we just check whether we hit a variable attribute or not.

Test added to verify it works. Tests are great.